+++ src/Instance/InstanceDetails/InstanceDescriptiveView/utils.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, screen } from '@testing-library/react';
+import { render, screen } from '@folio/jest-config-stripes/testing-library/react';
 import '../../../../test/jest/__mock__';
 import { formatLanguages } from './utils';