+++ src/edit/noteFields.js
@@ -17,6 +17,14 @@ import {
   Label,
 } from '@folio/stripes/components';
 
+import {
+  validateFieldLength
+} from '../utils';
+
+import {
+  NOTE_CHARS_MAX_LENGTH
+} from '../constants';
+
 const NoteFields = props => {
   const { formatMessage } = useIntl();
 
@@ -37,6 +45,8 @@ const NoteFields = props => {
   const isNoteTypeRequired = requiredFields.some(field => field === noteTypeIdField);
   const isNoteRequired = requiredFields.some(field => field === 'note');
   const isStaffOnlyRequired = requiredFields.some(field => field === 'staffOnly');
+
+  const validate = value => validateFieldLength(value, NOTE_CHARS_MAX_LENGTH);
 
   const headLabels = (
     <Row>
@@ -78,6 +88,7 @@ const NoteFields = props => {
           rows={1}
           disabled={!canEdit}
           required={isNoteRequired}
+          validate={validate}
         />
       </Col>
       <Col xs={3} lg={2}>