+++ src/hooks/useBoundWithTitlesByHrids/useBoundWithTitlesByHrids.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { renderHook } from '@testing-library/react-hooks';
+import { renderHook } from '@folio/jest-config-stripes/testing-library/react-hooks';
 
 import '../../../test/jest/__mock__';