+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change history for ui-inventory
 
-## 9.5.0 IN PROGRESS
+## 10.0.0 IN PROGRESS
 * Added a new option in the "Actions" dropdown within the Inventory app search page for "+New MARC Bib Record". Refs UIIN-2356.
 * Change the url for the "New MARC Bib Record" page. Refs UIIN-2380.
 * Avoid private paths in stripes-core imports. Refs UIIN-2367.
@@ -16,6 +16,10 @@
 * Item Create/Edit screens: Repeatable field trashcan is not aligned with the data row. Fixes UIIN-2374.
 * Chronology not displayed in receiving history. Fixes UIIN-2411.
 * Also support `circulation` `14.0`. Refs UIIN-2412.
+* Instance/Holdings/Item notes, administrative notes character limit to 32K. Refs UIIN-2354.
+* Import testing-library deps from `@folio/jest-config-stripes`. Refs UIIN-2427.
+* Bump zustand to v4. Refs UIIN-2353.
+* Fix the `records-editor/records` request by adding `_actionType`. Refs UIIN-2431.
 
 ## [9.4.5](https://github.com/folio-org/ui-inventory/tree/v9.4.5) (2023-04-03)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.4.4...v9.4.5)